+++ src/App.jsx
@@ -19,6 +19,7 @@ import PostListPage from "./pages/PostListPage";
 import SearchPage from "./pages/SearchPage";
 import PostDetailPage from "./pages/PostDetailPage";
 import LoginPage from "./pages/LoginPage";
+import NewProfilePage from "./pages/NewProfilePage";
 
 function App() {
   return (
@@ -36,6 +37,7 @@ function App() {
         <Route path="/signup" element={<SignUpPage />} />
         <Route path="/profile" element={<ProfilePage />} />
         <Route path="/profile/edit" element={<EditProfilePage />} />
+        <Route path="/profile/new" element={<NewProfilePage />} />
         <Route path="/post/upload" element={<PostUploadPage />} />
         <Route path="/product/add" element={<AddProductPage />} />
         <Route path="/follower" element={<FollowerListPage />} />
